split cache configuration and hazelcast configu
