+++ src/main/java/de/tum/cit/aet/artemis/core/config/EurekaClientConfiguration.java
@@ -29,6 +29,7 @@ import org.springframework.web.client.RestClient;
  */
 @Profile({ PROFILE_CORE, PROFILE_BUILDAGENT })
 @Configuration
+// TODO: check if we can make this class lazy
 @Lazy(value = false)
 public class EurekaClientConfiguration {
